+++ packages/solid-query/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.26](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.25...@wundergraph/solid-query@0.4.26) (2023-05-05)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
+## [0.4.25](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.24...@wundergraph/solid-query@0.4.25) (2023-05-04)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
+## [0.4.24](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.23...@wundergraph/solid-query@0.4.24) (2023-05-03)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
+## [0.4.23](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.22...@wundergraph/solid-query@0.4.23) (2023-05-03)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
 ## [0.4.22](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.21...@wundergraph/solid-query@0.4.22) (2023-04-27)
 
 **Note:** Version bump only for package @wundergraph/solid-query